move time_linearizer to extra.optimization.helpers [pr]
no longer used in tinygrad
